create status/otlp subpackage
